+++ tests/Microsoft.DotNet.Docker.Tests/CommonImageTests.cs
@@ -1,0 +1,70 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using System;
+using System.Collections.Generic;
+using System.Diagnostics.CodeAnalysis;
+using System.Linq;
+using Xunit;
+using Xunit.Abstractions;
+
+namespace Microsoft.DotNet.Docker.Tests
+{
+    public class CommonImageTests
+    {
+        private readonly DockerHelper _dockerHelper;
+        private readonly ITestOutputHelper _outputHelper;
+
+        public CommonImageTests(ITestOutputHelper outputHelper)
+        {
+            _dockerHelper = new DockerHelper(outputHelper);
+            _outputHelper = outputHelper;
+        }
+
+        public static IEnumerable<object[]> GetImageData()
+        {
+            return TestData.GetImageData()
+                // Filter the image data down to avoid duplicates caused by variations in their SDK OS.
+                .Distinct(new DefaultImageDataEqualityComparer())
+                .Select(imageData => new object[] { imageData });
+        }
+
+        [Theory]
+        [MemberData(nameof(GetImageData))]
+        public void VerifyImage_InsecureFilesCheck(ImageData imageData)
+        {
+            if (imageData.Version < new Version("3.1") || !DockerHelper.IsLinuxContainerModeEnabled ||
+                (imageData.OS.Contains("alpine") && imageData.IsArm))
+            {
+                return;
+            }
+
+            string worldWritableDirectoriesWithoutStickyBitCmd = @"find / -xdev -type d \( -perm -0002 -a ! -perm -1000 \)";
+            string worldWritableFilesCmd = "find / -xdev -type f -perm -o+w";
+            string noUserOrGroupFilesCmd;
+            if (imageData.OS.Contains("alpine"))
+            {
+                // BusyBox in Alpine doesn't support the more convenient -nouser and -nogroup options for the find command
+                noUserOrGroupFilesCmd = @"find / -xdev -exec stat -c %U-%n {} \+ | { grep ^UNKNOWN || true; }";
+            }
+            else
+            {
+                noUserOrGroupFilesCmd = @"find / -xdev \( -nouser -o -nogroup \)";
+            }
+
+            string command = $"/bin/sh -c \"{worldWritableDirectoriesWithoutStickyBitCmd} && {worldWritableFilesCmd} && {noUserOrGroupFilesCmd}\"";
+
+            foreach (DotNetImageType imageType in Enum.GetValues(typeof(DotNetImageType)))
+            {
+                string output = _dockerHelper.Run(
+                    image: imageData.GetImage(imageType, _dockerHelper),
+                    name: imageData.GetIdentifier($"InsecureFiles-{imageType}"),
+                    command: command
+                );
+
+                Assert.Empty(output);
+            }
+        }
+    }
+}
